Use picovision palette mode

+++ picovision/Display.h
@@ -3,7 +3,7 @@
 
 void init_display();
 
-void write_display(int x, int y, int count, uint16_t *data);
+void write_display(int x, int y, int count, uint8_t *data);
 void update_display();
 
 bool display_render_needed();
